Day 5 Solved
So I cheated a bit on this one.
I wanted the points so I formatted the input a bit to have them stack horizontally so I could read it quickly.
The next thing I might have to look at is how I'd read the data as it is. But for now, this works! :joy: